// src/openapi/client/models/GetUserFavorites200ResponseItemsInner.ts
export interface GetUserFavorites200ResponseItemsInner {
    /**
     * 
     * @type {number}
     * @memberof GetUserFavorites200ResponseItemsInner
     */
    id?: number;
    /**
     * 
     * @type {number}
     * @memberof GetUserFavorites200ResponseItemsInner
     */
    advertisementId?: number;
    /**
     * 
     * @type {string}
     * @memberof GetUserFavorites200ResponseItemsInner
     */
    brand?: string;
    /**
     * 
     * @type {string}
     * @memberof GetUserFavorites200ResponseItemsInner
     */
    model?: string;
    /**
     * 
     * @type {string}
     * @memberof GetUserFavorites200ResponseItemsInner
     */
    price?: string;
    /**
     * 
     * @type {string}
     * @memberof GetUserFavorites200ResponseItemsInner
     */
    currency?: string;
    /**
     * 
     * @type {number}
     * @memberof GetUserFavorites200ResponseItemsInner
     */
    releaseYear?: number;
    /**
     * 
     * @type {string}
     * @memberof GetUserFavorites200ResponseItemsInner
     */
    region?: string;
    /**
     * 
     * @type {Date}
     * @memberof GetUserFavorites200ResponseItemsInner
     */
    createdAt?: Date;
    /**
     * 
     * @type {Array<string>}
     * @memberof GetUserFavorites200ResponseItemsInner
     */
    images?: Array<string>;
}